requirements/development.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.4
django-autoslug==1.9.9
django-axes==6.1.1
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-dirtyfields==1.9.2
django-otp==1.3.0
django-phonenumber-field==7.1.0
django-simple-history==3.3.0
django-versatileimagefield==2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
multidict==6.0.4
phonenumbers==8.13.19
Pillow==10.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
sqlparse==0.4.4
twilio==8.11.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
